+++ packages/cli-v3/e2e/fixtures/config-invalid/trigger.config.js
@@ -1,0 +1,4 @@
+export const config = {
+  // 'project' field is mandatory
+  triggerDirectories: [],
+};