+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # internal-platform
 
+## 3.0.0-beta.44
+
+### Patch Changes
+
+- 39885a427: v3: fix missing init output in task run function when no middleware is defined
+
+## 3.0.0-beta.43
+
+### Patch Changes
+
+- 34ca7667d: v3: Include presigned urls for downloading large payloads and outputs when using runs.retrieve
+
 ## 3.0.0-beta.42
 
 ## 3.0.0-beta.41